+++ lib/screen/create group/create_group.dart
@@ -1,5 +1,5 @@
 import 'package:family_plus/models/user_model.dart';
-import 'package:family_plus/screen/home_screen.dart';
+import 'package:family_plus/screen/home/home_screen.dart';
 import 'package:family_plus/services/db_future.dart';
 import 'package:family_plus/states/current_user.dart';
 import 'package:firebase_auth/firebase_auth.dart';
