+++ lib/screen/login/login_screen.dart
@@ -1,5 +1,5 @@
-import 'package:family_plus/screen/home_screen.dart';
-import 'package:family_plus/screen/nav_screen.dart';
+import 'package:family_plus/screen/home/home_screen.dart';
+import 'package:family_plus/screen/navigation/nav_screen.dart';
 import 'package:family_plus/screen/signup/signup_screen.dart';
 import 'package:family_plus/services/auth_state_changes.dart';
 import 'package:family_plus/states/current_user.dart';
