+++ lib/screen/navigation/nav_screen.dart
@@ -1,9 +1,9 @@
-import 'package:family_plus/screen/add_todo.dart';
-import 'package:family_plus/screen/home_screen.dart';
+import 'package:family_plus/screen/todo/add_todo.dart';
+import 'package:family_plus/screen/home/home_screen.dart';
 import 'package:family_plus/screen/login/login_screen.dart';
-import 'package:family_plus/screen/logout_page.dart';
-import 'package:family_plus/screen/reward_page.dart';
-import 'package:family_plus/screen/todo_list.dart';
+import 'package:family_plus/screen/profile%20logout/logout_page.dart';
+import 'package:family_plus/screen/reward/reward_page.dart';
+import 'package:family_plus/screen/todo/todo_list.dart';
 import 'package:flutter/material.dart';
 // import 'package:flutter/matescreen/rial.dart';
 import 'package:persistent_bottom_nav_bar/persistent_tab_view.dart';
