+++ lib/screen/todo/todo_list.dart
@@ -1,5 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:family_plus/screen/view_data.dart';
+import 'package:family_plus/screen/todo/view_data.dart';
 import 'package:family_plus/widgets/todo_card.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
